Add update procedure for multi-instance
